добавил addData и addFormHandlers
